requirements.txt
Flask
Flask-SQLAlchemy
waitress
PyJWT
WeasyPrint
